refactor: typesフォルダをsrcフォルダの中に移動
