+++ src/utils/actionPost.ts
@@ -1,24 +1,18 @@
 import { getUuidFromCookie } from "@/actions/users";
-import { Post } from "../../types/post";
+import { Post } from "../types/post";
 import axios from "axios";
 
 // 投稿一覧を取得する関数
-export const getPosts = async (
-  page: number,
-  limit: number
-): Promise<Post[]> => {
+export const getPosts = async (page: number, limit: number): Promise<Post[]> => {
   try {
     const apiUrl = process.env.NEXT_PUBLIC_API_URL;
 
-    const response = await fetch(
-      `${apiUrl}/timeline/post?page=${page}&limit=${limit}`,
-      {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      }
-    );
+    const response = await fetch(`${apiUrl}/timeline/post?page=${page}&limit=${limit}`, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    });
 
     if (!response.ok) {
       throw new Error("Failed to fetch posts");
@@ -65,11 +59,7 @@ export const createPost = async (content: string): Promise<Post | null> => {
   }
 };
 
-export const sendReaction = async (
-  postId: string,
-  userId: string,
-  emoji: string
-) => {
+export const sendReaction = async (postId: string, userId: string, emoji: string) => {
   const apiUrl = process.env.NEXT_PUBLIC_API_URL;
   console.log("sendReaction", postId, userId, emoji);
   const response = await fetch(`${apiUrl}/timeline/add-reaction`, {
